[Update] clone libft and modified Makefile and somefiles

+++ srcs/print_helper.c
@@ -22,7 +22,7 @@ size_t	get_padding_len(size_t print_len, t_info *info)
 	size_t	padding_len;
 
 	padding_len = 0;
-	if (info->min_width >= 0 && info->min_width > print_len)
+	if (info->min_width >= 0 && (size_t)info->min_width > print_len)
 		padding_len = info->min_width - print_len;
 	return (padding_len);
 }
@@ -35,7 +35,7 @@ size_t	get_print_len(char *s, t_info *info)
 		print_len = ft_strlen(NULL_STR);
 	else
 		print_len = ft_strlen(s);
-	if (info->precision >= 0 && print_len > info->precision)
+	if (info->precision >= 0 && print_len > (size_t)info->precision)
 		print_len = info->precision;
 	return (print_len);
 }
