+++ srcs/print_helper2.c
@@ -5,7 +5,7 @@ size_t	get_print_len_n(char *nbr, t_info *info)
 	size_t	precision_len;
 
 	precision_len = ft_strlen(nbr);
-	if (info->precision >= 0 && info->precision > precision_len)
+	if (info->precision >= 0 && (size_t)info->precision > precision_len)
 		precision_len = info->precision;
 	return (precision_len);
 }
@@ -17,7 +17,7 @@ size_t	get_zero_pad_len(char *nbr, t_info *info)
 
 	zero_pad_len = 0;
 	nbr_len = ft_strlen(nbr);
-	if (info->precision >= 0 && info->precision > nbr_len)
+	if (info->precision >= 0 && (size_t)info->precision > nbr_len)
 		zero_pad_len = info->precision - nbr_len;
 	return (zero_pad_len);
 }
@@ -27,7 +27,7 @@ size_t	get_print_len_hex(unsigned int n, char *nbr, t_info *info)
 	size_t	print_len;
 
 	print_len = ft_strlen(nbr);
-	if (info->precision >= 0 && info->precision > print_len)
+	if (info->precision >= 0 && (size_t)info->precision > print_len)
 		print_len = info->precision;
 	if (n != 0 && info->alternate)
 		print_len += 2;
@@ -39,7 +39,7 @@ size_t	get_print_len_di(int n, char *nbr, t_info *info)
 	size_t	print_len;
 
 	print_len = ft_strlen(nbr);
-	if (info->precision >= 0 && info->precision > print_len)
+	if (info->precision >= 0 && (size_t)info->precision > print_len)
 		print_len = info->precision;
 	if ((info->sign_space || info->sign) && n >= 0)
 		print_len++;
